Implement auto-login for Arena V3 integrations and improve session handling in workspace layout
