+++ apps/sim/app/workspace/[workspaceId]/layout.tsx
@@ -1,4 +1,5 @@
 import { Suspense } from 'react'
+import { cookies } from 'next/headers'
 import { redirect } from 'next/navigation'
 import { ToastProvider } from '@/components/emcn'
 import { getSession } from '@/lib/auth'
@@ -25,10 +26,16 @@ export default function WorkspaceLayout({ children }: { children: React.ReactNod
 async function WorkspaceLayoutInner({ children }: { children: React.ReactNode }) {
   const session = await getSession()
   if (!session?.user) {
-    redirect('/login')
+    //this logic is for doing auto login
+    const cookieStore = await cookies()
+    const hasEmailCookie = !!cookieStore.get('email')?.value
+    if (!hasEmailCookie) {
+      redirect('/login')
+    }
+    //--------------
   }
   // The organization plugin is conditionally spread so TS can't infer activeOrganizationId on the base session type.
-  const orgId = (session.session as { activeOrganizationId?: string } | null)?.activeOrganizationId
+  const orgId = (session?.session as { activeOrganizationId?: string } | null)?.activeOrganizationId
   const initialOrgSettings = orgId ? await getOrgWhitelabelSettings(orgId) : null
 
   return (
