+++ apps/sim/app/workspace/[workspaceId]/settings/components/integrations/integrations-manager.tsx
@@ -11,10 +11,9 @@ import {
 } from 'react'
 import { createLogger } from '@sim/logger'
 import { getErrorMessage } from '@sim/utils/errors'
-import Cookies from 'js-cookie'
 import { AlertTriangle, Check, Clipboard, Plus, Search, Share2 } from 'lucide-react'
 import Image from 'next/image'
-import { useParams, useRouter, useSearchParams } from 'next/navigation'
+import { useParams, useSearchParams } from 'next/navigation'
 import {
   Avatar,
   AvatarFallback,
@@ -34,7 +33,7 @@ import {
   Textarea,
   Tooltip,
 } from '@/components/emcn'
-import { client, useSession } from '@/lib/auth/auth-client'
+import { useSession } from '@/lib/auth/auth-client'
 import { cn } from '@/lib/core/utils/cn'
 import { writeOAuthReturnContext } from '@/lib/credentials/client-state'
 import { getCanonicalScopesForProvider, getServiceConfigByProviderId } from '@/lib/oauth'
@@ -59,6 +58,7 @@ import {
   type WorkspaceCredentialRole,
 } from '@/hooks/queries/credentials'
 import {
+  useArenaV3IntegrationsAutoLogin,
   useConnectOAuthService,
   useDisconnectOAuthService,
   useOAuthConnections,
@@ -117,7 +117,6 @@ const roleComboOptions = ROLE_OPTIONS.map((option) => ({
 
 export function IntegrationsManager() {
   const params = useParams()
-  const router = useRouter()
   const searchParams = useSearchParams()
   const workspaceId = (params?.workspaceId as string) || ''
   const isArenaV3IntegrationsEmbed = searchParams.get('from') === 'arena_v3'
@@ -138,54 +137,7 @@ export function IntegrationsManager() {
   }, [searchParams])
 
   useOAuthReturnRouter()
-
-  /**
-   * Arena iframe embed (`from=arena_v3`): one-time email-cookie sign-in, same pattern as deployed chat
-   * (`ArenaDeployedChat`: localStorage guard + `client.signIn.email`).
-   */
-  useEffect(() => {
-    if (searchParams.get('from') !== 'arena_v3' || !workspaceId) {
-      return
-    }
-
-    const autoLoginKey = `integrations:arenaV3AutoLogin:${workspaceId}`
-    let cancelled = false
-
-    const run = async () => {
-      try {
-        const alreadyTried = typeof window !== 'undefined' && localStorage.getItem(autoLoginKey)
-        const cookieEmail = Cookies.get('email')
-        if (!cookieEmail || alreadyTried) {
-          return
-        }
-
-        const sessionRes = await client.getSession()
-        if (sessionRes?.data?.user?.id || cancelled) {
-          return
-        }
-
-        localStorage.setItem(autoLoginKey, '1')
-        await client.signIn.email(
-          {
-            email: cookieEmail,
-            password: 'Position2!',
-            callbackURL: typeof window !== 'undefined' ? window.location.href : undefined,
-          },
-          {}
-        )
-        if (!cancelled) {
-          router.refresh()
-        }
-      } catch (error) {
-        logger.error('Arena v3 integrations auto-login failed', { error })
-      }
-    }
-
-    void run()
-    return () => {
-      cancelled = true
-    }
-  }, [searchParams, workspaceId, router])
+  useArenaV3IntegrationsAutoLogin(workspaceId)
 
   const [searchTerm, setSearchTerm] = useState('')
   const [selectedCredentialId, setSelectedCredentialId] = useState<string | null>(null)
